chore: refactor operations routes
